o cpu     - corrected initial state of break and interrupt disable flags on 	startup     - result.InstructionData now set as soon as possible in all 	instances     - corrected RTI implementation     - corrected RTS implementation     - implemented some undocumented instructions: 	0x0c skw 	0xb3 lax 	0xc7 dcp 	0x4b asr     - a homebrew version of pacman (pac-man_4k.bin) now works (up to a 	point; not sure if bug is in the implementation of the new 	instructions or is a video issue)     - new effect classification "Interrupt" for BRK and RTI     - instructions_gen program is now more robust
o disassembly
    - added validity check to ParseMemory() function
    - Dump() now uses strings.Builder

o result.Instruction
    - corrected GetString() function
	- function would panic when in videostep mode
	- bug most probably introduced in 58f452cf52343eb27fb495b3b0f6e6675806203f
    - improved output of IsValid()
	- disassembly of problem line is now included as part of the
	    error string

+++ hardware/vcs.go
@@ -13,7 +13,7 @@ import (
 
 // AddressReset is the address where the reset address is stored
 // - used by VCS.Reset() and Disassembly module
-const AddressReset = 0xfffc
+const AddressReset = uint16(0xfffc)
 
 // AddressIRQ is the address where the interrupt address is stored
 const AddressIRQ = 0xfffe
